Refactor sidebar item
